language-bindings: simplify versions
"GTK+ 2" and "GTK+ 3" are much easier to maintain than more precise
versions like "2.24" and "3.10".

With only "GTK+ 3", the page looks up-to-date, while with "3.10" the
page is outdated.

Some GTK+ 3 bindings support the latest 3.x versions, but others don't.
For the D language for example, the latest supported version is 3.10, so
updating the versions in the table to 3.14 is impossible, unless another
column is added.