Add package manager info to install page
Finding the Debian package for gtk3 in particular is a bit of a pain. For instance, searching "gtk" or "gtk3" in https://packages.ubuntu.com/ doesn't yield the correct package.

Since this page was the first hit I searched "installing gtk3", I'm proposing to add basic instructions to find the correct package here.